Delete script and deleteDoc

+++ db/dbhelper.js
@@ -65,30 +65,48 @@ var dbFunc = {
 	getScripts: function(username, res) {
 		Model.user.findOne({'username': username}).populate('scripts').exec(function (err, found) {
 			if(err){
-				console.log('error in fetching tasks', err);
+				console.log('error in fetching scripts', err);
 			}
 			console.log('these are the found scripts for harish', found.scripts)
 			res.send(found.scripts);
 		});
 	},
 
-  addDoc: function(data, res) {
-  	var newDoc = new Model.doctor(data);
+  addDoc: function(data, res, next) {
+		console.log("addDoc called with", data);
+  	var newDoc = new Model.doctor(data.doc);
   	newDoc.save(function(err) {
   		if (err) {
   			console.log(err);
   		}
-  		console.log("Doctor added!");
-  		res.send(newDoc);
+  		Model.user.update({"username": data.username}, {$push:{"doctors": newDoc}}, function(err){
+				if(err){
+					next(new Error("doctor added to user model"))
+				}
+				res.send(newDoc)
+			})
   	});
   },
 
-  getDocs: function(target, res) {
-    Model.doctor.find({}, function(err, docs) {
-      console.log(docs);
-      res.send(docs);
-    });
+  getDocs: function(username, res, next) {
+		Model.user.findOne({"username": username}).populate('doctors').exec(function(err, user){
+			if(err){
+				next(new Error(err));
+			}
+			console.log("?!?!", user);
+			res.send(user.doctors)
+		})
   },
+
+	deleteDoc: function(id, res, next) {
+		Model.doctor.remove({"_id": id}, function(err){
+			if(err){
+				next("reminder not deleted", err);
+			}
+			next("doctor deleted");
+		})
+  },
+
 
 	/* AUTHENTICATION FUNCTIONS */
 
@@ -142,10 +160,6 @@ var dbFunc = {
 						var token = jwt.encode(user, 'secret'); //create new token
 						console.log('this is token',token);
 						var resultData = {"token": token, "user": {"id": user._id, "username": user.username}}
-						console.log('to be sent', resultData);
-						console.log(res);
-						// res.send({"token": token, "user": {"id": user._id, "username": user.username}});
-						// next(JSON.stringify({"token": token, "user": {"id": user._id, "username": user.username}}));
 	          res.status(201).send(resultData); //send new token and user object
 					}
 				});
@@ -160,7 +174,6 @@ var dbFunc = {
 	  }
 	  else {
 	    var user = jwt.decode(token, 'secret');
-	    console.log("Decoded user:", user);
 	    Model.user.find(user, function(err, user){
 	    	if(err){
 	    		next("Error: ", error);
@@ -224,7 +237,9 @@ deleteReminder: function(reminderID, next) {
 		if(err){
 			next("reminder not deleted", err);
 		}
-		reminderManager.deleteJob(reminderID.toString());
+		if(reminderManager.exists(reminderID.toString())) {
+			reminderManager.deleteJob(reminderID.toString())
+		}
 		next("reminder deleted");
 	})
 },
